requirements.txt
streamlit==1.38.0
crewai==0.61.1
google-generativeai==0.8.3
langchain-google-genai==2.0.0
langchain-community==0.2.12
faiss-cpu==1.8.0
pydantic==2.8.0
python-dotenv==1.0.1
